Add core namespace that's got all the errors
