+++ core/src/main/scala/io/mediachain/core/Error.scala
@@ -1,9 +1,14 @@
-package io.mediachain
-
+package io.mediachain.core
 import gremlin.scala.Vertex
 
-sealed abstract class GraphError
+sealed abstract class TranslationError
+object TranslationError {
+  case class ParsingFailed(underlying: Throwable) extends TranslationError
+  case class ResourceNotReadable(underlying: Throwable) extends TranslationError
+  case class InvalidFormat() extends TranslationError
+}
 
+sealed abstract class GraphError
 object GraphError {
   case class TooManyRawBlobsError(vertex: Vertex) extends GraphError
   case class AuthorNotFoundError(vertex: Vertex) extends GraphError
@@ -16,9 +21,7 @@ object GraphError {
   case class SubtreeError() extends GraphError
 }
 
-
 sealed abstract class MediachainError
-
 object MediachainError {
   case class Graph(error: GraphError) extends MediachainError
 }
