--- translation_engine/src/main/scala/io/mediachain/translation/TranslationError.scala
@@ -1,9 +1,0 @@
-package io.mediachain.translation
-
-sealed abstract class TranslationError
-
-object TranslationError {
-  case class ParsingFailed(underlying: Throwable) extends TranslationError
-  case class ResourceNotReadable(underlying: Throwable) extends TranslationError
-  case class InvalidFormat() extends TranslationError
-}
